EBAC-BackendPython/Parte-3/bookstore/pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "Bookstore API"
authors = ["franc <franckallyson@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
django-rest-framework = "^0.1.0"
pytest-django = "^4.5.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
factory-boy = "^3.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
